requirements.txt
numpy
soundfile
opencv-python
pillow
pygame
fpstimer
ffmpeg
moviepy
rich
click